Fix @jaredly's bug.
I introduced this when I first wrote the watch crawler. The issue was
that when you're watch mode you don't need to make-world (since you know
you just built the world in the previous run), so you can skip a bunch
of stuff. The issue is that with bytecode/native compilation we still
need to crawl the dependency tree to figure out all of the cma/cmxa
library files to link into the final binary. The code that did that
crawling was wrong and didn't take into account the allowed-build-kinds.

The fix is to check that the library is buildable for the current target.