WIP allow ppxes inside project source.
This works for native/bytecode, still need to make the changes to JS
compilation. This works by adding a new `type` to the `sources` key
called "ppx" which tells the compiler that it's a separate thing.
Then you add an entry to the `entries` with a `type` "ppx" as well.
Then other entries can add `ppx` and depend on that module.