Make Belt.List work and add C code.
I didn't do a pass on all externals yet, so this wasn't linking.
Introduced C code to prevent a compiler optimization which would do
inlining things to lists created like `hd :: tl` because it assumes
immutability. Luckily we can trick the compiler using an external which
does the same thing, and now we have a mutable list! We need to use it
sparingly.