Fix bs.deriving to behave like JS's version.
Before I used to expose the record type to the rest of the file, so
you could construct it with a record literal. That was so that the
constructor could work (it needs to make the record with a record
literal).
That wasn't totally correct, so some code would compile to native that
wouldn't work in JS.
I updated the belt_bsppx to add a module signature which hides the
implementation of the type (making it abstract). That should be enough
so that you can't construct it using a literal, even inside the same
file. This is done in the same pass, should be pretty cheap *crosses
fingers*.